# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NovaRewards API",
    "version": "1.0.0",
    "description": "REST API for the Nova Rewards blockchain loyalty platform. User endpoints require a Bearer JWT from `POST /auth/login`. Merchant endpoints require an `x-api-key` header.",
    "contact": {
      "name": "Nova Rewards Team"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local development"
    },
    {
      "url": "https://api.novarewards.io/api",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from `POST /auth/login`."
      },
      "merchantApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Merchant API key returned on registration (shown once)."
      }
    },
...
  },
  "paths": {
...
    "/campaigns": {
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Create campaign and register on-chain",
        "security": [
          {
            "merchantApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "rewardRate",
                  "startDate",
                  "endDate"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "rewardRate": {
                    "type": "number",
                    "example": 1.5
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campaign created"
          },
          "400": {
            "description": "Validation error"
          },
          "502": {
            "description": "On-chain error"
          }
        }
      },
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "List campaigns for authenticated merchant (cached 60s)",
        "security": [
          {
            "merchantApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign list"
          }
        }
      }
    },
    "/campaigns/{id}": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Get campaign by ID",
        "security": [
          {
            "merchantApiKey": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign"
          },
          "403": {
            "description": "Not own campaign"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Update campaign and push on-chain",
        "security": [
          {
            "merchantApiKey": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "rewardRate": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated campaign"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Not confirmed on-chain"
          },
          "502": {
            "description": "Chain error"
          }
        }
      },
      "delete": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Pause on-chain then soft-delete",
        "security": [
          {
            "merchantApiKey": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "502": {
            "description": "Chain error"
          }
        }
      }
    },
...
    "/drops/eligible": {
      "get": {
        "tags": [
          "Drops"
        ],
        "summary": "List drops the authenticated user qualifies for",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Eligible drops"
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
    "/drops/{id}/claim": {
      "post": {
        "tags": [
          "Drops"
        ],
        "summary": "Claim a drop",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Drop claimed"
          },
          "400": {
            "description": "Already claimed or not eligible"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "404": {
            "description": "Drop not found"
          }
        }
      }
    },
...
    "/admin/campaigns/{id}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Force-delete a campaign (admin override)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "403": {
            "description": "Admin required"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  }
}